+++ .toys/.toys.rb
@@ -25,15 +25,3 @@
 # This is to support development of Toys itself. Most Toys files should not
 # include this.
 ::Kernel.exec(::File.join(context_directory, "toys-dev"), *::ARGV) unless ::ENV["TOYS_DEV"]
-
-include :gems, suppress_confirm: true
-gem "highline", "~> 2.0"
-gem "minitest", "~> 5.13"
-gem "minitest-focus", "~> 1.1"
-gem "minitest-rg", "~> 5.2"
-gem "rake", "~> 13.0"
-gem "rdoc", "~> 6.1.2"
-gem "redcarpet", "~> 3.5" unless ::RUBY_PLATFORM == "java"
-gem "rspec", "~> 3.9"
-gem "rubocop", "~> 0.78.0"
-gem "yard", "~> 0.9.22"