+++ .toys/yardoc.rb
@@ -24,12 +24,13 @@
 desc "Generates yardoc for both gems"
 
 include :terminal
+include :exec
 
 def handle_gem(gem_name)
   puts("**** Generating Yardoc for #{gem_name}...", :bold, :cyan)
   ::Dir.chdir(::File.join(context_directory, gem_name)) do
-    status = cli.child.add_config_path(".toys.rb").run("yardoc")
-    exit(status) unless status.zero?
+    result = exec_separate_tool("yardoc")
+    exit(result.exit_code) unless result.success?
   end
 end