+++ toys/lib/toys.rb
@@ -21,6 +21,20 @@
 # IN THE SOFTWARE.
 ;
 
+require "toys/version"
+
+# Add toys-core to the load path. The Toys debug scripts will set this
+# environment variable explicitly, but in production, we get it from rubygems.
+# We prepend to $LOAD_PATH directly rather than calling Kernel.gem, so that we
+# don't get clobbered in case someone sets up bundler later.
+::ENV["TOYS_CORE_LIB_PATH"] ||= begin
+  dep = Gem::Dependency.new("toys-core", "= #{::Toys::VERSION}")
+  dep.to_spec.full_require_paths.first
+end
+$LOAD_PATH.unshift(::ENV["TOYS_CORE_LIB_PATH"])
+
+require "toys-core"
+
 ##
 # Toys is a configurable command line tool. Write commands in config files
 # using a simple DSL, and Toys will provide the command line executable and
@@ -45,17 +59,6 @@ module Toys
   module Templates; end
 end
 
-require "toys/version"
+::Toys.executable_path = ::Toys::EXECUTABLE_PATH
 
-# Add toys-core to the load path. The Toys debug scripts will set this
-# environment variable explicitly, but in production, we get it from rubygems.
-# We prepend to $LOAD_PATH directly rather than calling Kernel.gem, so that we
-# don't get clobbered in case someone sets up bundler later.
-::ENV["TOYS_CORE_LIB_PATH"] ||= begin
-  dep = Gem::Dependency.new("toys-core", "= #{::Toys::VERSION}")
-  dep.to_spec.full_require_paths.first
-end
-$LOAD_PATH.unshift(::ENV["TOYS_CORE_LIB_PATH"])
-
-require "toys-core"
 require "toys/standard_cli"