+++ toys/lib/toys/templates/minitest.rb
@@ -66,6 +66,11 @@ module Toys
       #     files to load. Defaults to {DEFAULT_FILES}.
       # @param warnings [Boolean] If true, runs tests with Ruby warnings.
       #     Defaults to true.
+      # @param bundler [Boolean,Hash] If `false` (the default), bundler is not
+      #     enabled for this tool. If `true` or a Hash of options, bundler is
+      #     enabled. See the documentation for the
+      #     [bundler mixin](https://dazuma.github.io/toys/gems/toys-core/latest/Toys/StandardMixins/Bundler)
+      #     for information on available options.
       #
       def initialize(name: nil,
                      gem_version: nil,
@@ -73,7 +78,8 @@ module Toys
                      files: nil,
                      seed: nil,
                      verbose: false,
-                     warnings: true)
+                     warnings: true,
+                     bundler: nil)
         @name = name || DEFAULT_TOOL_NAME
         @gem_version = gem_version || DEFAULT_GEM_VERSION_REQUIREMENTS
         @libs = libs || DEFAULT_LIBS
@@ -81,6 +87,7 @@ module Toys
         @seed = seed
         @verbose = verbose
         @warnings = warnings
+        self.bundler = bundler
       end
 
       attr_accessor :name
@@ -91,12 +98,55 @@ module Toys
       attr_accessor :verbose
       attr_accessor :warnings
 
+      ##
+      # Activate bundler for this tool.
+      #
+      # See the documentation for the
+      # [bundler mixin](https://dazuma.github.io/toys/gems/toys-core/latest/Toys/StandardMixins/Bundler)
+      # for information on the options that can be passed.
+      #
+      # @param opts [keywords] Options for bundler
+      # @return [self]
+      #
+      def bundler(**opts)
+        @bundler_settings = opts
+        self
+      end
+
+      ##
+      # Set the bundler state and options for this tool.
+      #
+      # Pass `false` to disable bundler. Pass `true` or a hash of options to
+      # enable bundler. See the documentation for the
+      # [bundler mixin](https://dazuma.github.io/toys/gems/toys-core/latest/Toys/StandardMixins/Bundler)
+      # for information on the options that can be passed.
+      #
+      # @param opts [true,false,Hash] Whether bundler should be enabled for
+      #     this tool.
+      # @return [self]
+      #
+      def bundler=(opts)
+        @bundler_settings =
+          if opts && !opts.is_a?(::Hash)
+            {}
+          else
+            opts
+          end
+      end
+
+      ## @private
+      attr_reader :bundler_settings
+
       on_expand do |template|
         tool(template.name) do
           desc "Run minitest on the current project."
 
           include :exec
           include :gems
+
+          if template.bundler_settings
+            include :bundler, **template.bundler_settings
+          end
 
           flag :seed, "-s", "--seed SEED",
                default: template.seed, desc: "Sets random seed."