+++ toys/lib/toys/templates/yardoc.rb
@@ -80,6 +80,11 @@ module Toys
       # @param options [Array<String>] Additional options passed to YARD
       # @param stats_options [Array<String>] Additional options passed to YARD
       #     stats
+      # @param bundler [Boolean,Hash] If `false` (the default), bundler is not
+      #     enabled for this tool. If `true` or a Hash of options, bundler is
+      #     enabled. See the documentation for the
+      #     [bundler mixin](https://dazuma.github.io/toys/gems/toys-core/latest/Toys/StandardMixins/Bundler)
+      #     for information on available options.
       #
       def initialize(name: nil,
                      gem_version: nil,
@@ -99,7 +104,8 @@ module Toys
                      template_path: nil,
                      format: nil,
                      options: [],
-                     stats_options: [])
+                     stats_options: [],
+                     bundler: nil)
         @name = name || DEFAULT_TOOL_NAME
         @gem_version = gem_version || DEFAULT_GEM_VERSION_REQUIREMENTS
         @files = files
@@ -119,6 +125,7 @@ module Toys
         @format = format
         @options = options
         @stats_options = stats_options
+        self.bundler = bundler
       end
 
       attr_accessor :name
@@ -141,6 +148,45 @@ module Toys
       attr_accessor :options
       attr_accessor :stats_options
 
+      ##
+      # Activate bundler for this tool.
+      #
+      # See the documentation for the
+      # [bundler mixin](https://dazuma.github.io/toys/gems/toys-core/latest/Toys/StandardMixins/Bundler)
+      # for information on the options that can be passed.
+      #
+      # @param opts [keywords] Options for bundler
+      # @return [self]
+      #
+      def bundler(**opts)
+        @bundler_settings = opts
+        self
+      end
+
+      ##
+      # Set the bundler state and options for this tool.
+      #
+      # Pass `false` to disable bundler. Pass `true` or a hash of options to
+      # enable bundler. See the documentation for the
+      # [bundler mixin](https://dazuma.github.io/toys/gems/toys-core/latest/Toys/StandardMixins/Bundler)
+      # for information on the options that can be passed.
+      #
+      # @param opts [true,false,Hash] Whether bundler should be enabled for
+      #     this tool.
+      # @return [self]
+      #
+      def bundler=(opts)
+        @bundler_settings =
+          if opts && !opts.is_a?(::Hash)
+            {}
+          else
+            opts
+          end
+      end
+
+      ## @private
+      attr_reader :bundler_settings
+
       on_expand do |template|
         tool(template.name) do
           desc "Run yardoc on the current project."
@@ -156,6 +202,10 @@ module Toys
           include :exec
           include :terminal
           include :gems
+
+          if template.bundler_settings
+            include :bundler, **template.bundler_settings
+          end
 
           to_run do
             gem_requirements = Array(template.gem_version)